requirements.txt
flask
flask-cors
newspaper3k
llama-index
langchain
python-dotenv
youtube-transcript-api
openai
wolframalpha
google-search-results
wikipedia
google-api-python-client
transformers
sentence_transformers
huggingface_hub
configparser
replicate